restructuring the files
