+++ packages/contracts/src/executor/ConstructorTxExecutor.ts
@@ -1,18 +1,9 @@
 import { GenericSubstrateApi } from '@dedot/types';
-import {
-  assert,
-  assertFalse,
-  concatU8a,
-  hexToU8a,
-  isNull,
-  isUndefined,
-  isWasm,
-  u8aToHex,
-} from '@dedot/utils';
+import { assert, assertFalse, concatU8a, hexToU8a, isNull, isUndefined, isWasm, u8aToHex } from '@dedot/utils';
 import { ConstructorTxOptions, GenericConstructorTxCall } from '../types/index.js';
-import {DeployerExecutor} from "./DeployerExecutor.js";
+import { DeployerExecutor } from './abstract/index.js';
 
-export class ConstructorTxExecutor<ChainApi extends GenericSubstrateApi> extends DeployerExecutor <ChainApi> {
+export class ConstructorTxExecutor<ChainApi extends GenericSubstrateApi> extends DeployerExecutor<ChainApi> {
   doExecute(constructor: string) {
     const meta = this.findConstructorMeta(constructor);
     assert(meta, `Constructor message not found: ${constructor}`);
