+++ packages/contracts/src/executor/TxExecutor.ts
@@ -1,7 +1,7 @@
 import { GenericSubstrateApi } from '@dedot/types';
 import { assert, concatU8a, hexToU8a, u8aToHex } from '@dedot/utils';
 import { ContractTxOptions, GenericContractTxCall } from '../types/index.js';
-import { ContractExecutor } from './ContractExecutor.js';
+import { ContractExecutor } from './abstract/index.js';
 
 export class TxExecutor<ChainApi extends GenericSubstrateApi> extends ContractExecutor<ChainApi> {
   doExecute(message: string) {
