+++ packages/contracts/src/executor/abstract/ContractExecutor.ts
@@ -1,9 +1,9 @@
 import { ISubstrateClient } from '@dedot/api';
 import { AccountId32, AccountId32Like } from '@dedot/codecs';
 import { GenericSubstrateApi } from '@dedot/types';
-import { TypinkRegistry } from '../TypinkRegistry.js';
-import { ContractCallMessage, ExecutionOptions } from '../types/index.js';
-import { normalizeLabel } from '../utils.js';
+import { TypinkRegistry } from '../../TypinkRegistry.js';
+import { ContractCallMessage, ExecutionOptions } from '../../types/index.js';
+import { normalizeLabel } from '../../utils.js';
 import { Executor } from './Executor.js';
 
 export abstract class ContractExecutor<ChainApi extends GenericSubstrateApi> extends Executor<ChainApi> {
