+++ packages/contracts/src/executor/abstract/DeployerExecutor.ts
@@ -2,9 +2,9 @@ import { ISubstrateClient } from '@dedot/api';
 import { Hash } from '@dedot/codecs';
 import { GenericSubstrateApi } from '@dedot/types';
 import { HexString } from '@dedot/utils';
-import { TypinkRegistry } from '../TypinkRegistry.js';
-import { ContractConstructorMessage, ExecutionOptions } from '../types/index.js';
-import { normalizeLabel } from '../utils.js';
+import { TypinkRegistry } from '../../TypinkRegistry.js';
+import { ContractConstructorMessage, ExecutionOptions } from '../../types/index.js';
+import { normalizeLabel } from '../../utils.js';
 import { Executor } from './Executor.js';
 
 export abstract class DeployerExecutor<ChainApi extends GenericSubstrateApi> extends Executor<ChainApi> {
