+++ packages/contracts/src/executor/abstract/Executor.ts
@@ -1,8 +1,8 @@
 import { ISubstrateClient } from '@dedot/api';
 import { SubstrateApi } from '@dedot/api/chaintypes';
 import { GenericSubstrateApi, RpcVersion } from '@dedot/types';
-import { TypinkRegistry } from '../TypinkRegistry.js';
-import { ContractMessageArg, ContractMessage, ExecutionOptions, ContractMetadata } from '../types/index.js';
+import { TypinkRegistry } from '../../TypinkRegistry.js';
+import { ContractMessageArg, ContractMessage, ExecutionOptions, ContractMetadata } from '../../types';
 
 export abstract class Executor<ChainApi extends GenericSubstrateApi = SubstrateApi[RpcVersion]> {
   constructor(
