+++ packages/contracts/src/executor/index.ts
@@ -1,4 +1,4 @@
-export * from './Executor.js';
+export * from './abstract/index.js';
 export * from './TxExecutor.js';
 export * from './QueryExecutor.js';
 export * from './ConstructorTxExecutor.js';
